secondClient

+++ src/main/java/com/example/dao/CategoryDAO.java
@@ -3,6 +3,6 @@ package com.example.dao;
 import com.example.categories.CourseCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryDAO extends JpaRepository<CourseCategory, Integer> {
+public interface CategoryDAO extends JpaRepository<CourseCategory, Long> {
 
 }
